+++ src/enums/button-icon.ts
@@ -19,7 +19,7 @@ export enum ButtonIconEnum {
   /** 刷新/同步 */
   refresh = 'solar:refresh-broken',
   /** 未收藏 */
-  favorite = 'carbon:favorite',
+  favorite = 'heart',
   /** 已收藏 */
-  favoriteFilled = 'carbon:favorite-filled',
+  favoriteFilled = 'heart-filled',
 }
