+++ src/views/play-station/profile/components/game-card/index.vue
