+++ src/views/play-station/profile/components/game-list/components/index.ts
