+++ src/views/play-station/profile/components/game-list/components/sync-game-modal.vue
