+++ src/views/play-station/profile/components/game-list/index.vue
