+++ src/views/play-station/profile/components/profile-info/index.vue
