--- src/views/play-station/trophy/index.vue
@@ -1,11 +1,0 @@
-<template>
-  <div>PlayStationTrophyView</div>
-</template>
-
-<script lang="ts" setup>
-defineOptions({
-  name: 'PlayStationTrophyView',
-});
-</script>
-
-<style lang="scss" scoped></style>
